Added changes to support nodes of the following form
[:parent-tag {...}
  (([:tag1] [:tag2])
   [:tag3])]

equivalent to

[:parent-tag {...}
  [:tag1] [:tag2] [:tag3]]

This is useful when building up data from various functions where you dont want to wrap children in [:div] or other at each stage. This is consistent with behaviour in crate and hiccup.
Effectively seqs are looked-through for elements that can be rendered.

Minor change to slice removing "non-compliant" js/Array.prototype.slice